+++ ClosedXML/Excel/AutoFilters/XLFilterColumn.cs
@@ -8,9 +8,9 @@ namespace ClosedXML.Excel
     internal class XLFilterColumn : IXLFilterColumn
     {
         private readonly XLAutoFilter _autoFilter;
-        private readonly Int32 _column;
+        private readonly short _column;
 
-        public XLFilterColumn(XLAutoFilter autoFilter, Int32 column)
+        public XLFilterColumn(XLAutoFilter autoFilter, short column)
         {
             _autoFilter = autoFilter;
             _column = column;
